apps/api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.3
geoalchemy2==0.15.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.1.1
python-dotenv==1.0.1
web3==7.2.0
alembic==1.13.3
shapely==2.0.6
email-validator==2.2.0
python-multipart==0.0.18
